Moved test case creation to separate package
