+++ internal/testcase/gen_type_correct.go
@@ -1,4 +1,4 @@
-package gen
+package testcase
 
 import (
 	"go/ast"
@@ -7,7 +7,7 @@ import (
 )
 
 // CorrectTypeExpr corrects type expressions for imports
-func (g *Generator) CorrectTypeExpr(e ast.Expr, input *RecursionInput) ast.Expr { // nolint: funlen, gocyclo
+func (g *TestCase) CorrectTypeExpr(e ast.Expr, input *RecursionInput) ast.Expr { // nolint: funlen, gocyclo
 	switch t := e.(type) {
 	case *ast.ArrayType:
 		return &ast.ArrayType{
